feat(all): new loader-plugin-based remplate loader
This is a BREAKING CHANGE. The new default loader uses plugins to
System.js and Require.js which then delegate to the html import
capability. This allows Aurelia to use the module loader’s map/path
configuration for views, providing a more correct resolution
capability. The loader now returns a TemplateRegistryEntry instead of
just the raw template. This entry contains the template, its
dependencies and its id. It is expect that a view engine will also
provide the resources and factory at a later time. All requests for the
same template should return the same entry instance.

As a result of loader-based view loading, we have been able to remove
the baseUrl and baseViewUrl properties (which were horribly hacked
anyway).

Finally, this loader will expose itself as window.AureliaLoader for use
by the bootstrapper.