fix(all): remove web components and html imports
BREAKING CHANGE - Now all views will be loaded with the text loader by
default. This is included for system.js. If you are using require.js
you need to install its text plugin for this to work. If you are using
dojo, you need to set loader.textPluginName = “dojo/text” for this to
work. If you are using Polymer and/or wish to use HTMLImport for views
instead, you should install the html-import-template-loader plugin
during app bootstrap to restore view loading to using HTMLImport
technology.